requirements.txt
alembic==1.16.5
babel==2.17.0
blinker==1.9.0
Brotli==1.1.0
cffi==2.0.0
click==8.2.1
cryptography==46.0.2
cssselect2==0.8.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
fonttools==4.59.2
fpdf==1.7.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
packaging==25.0
pillow==11.3.0
pycparser==2.23
pydyf==0.11.0
PyMySQL==1.1.2
pyphen==0.17.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.0.1
zopfli==0.2.3.post1
